move product code to its own module to keep separate from testmonitor
