+++ nisystemlink/clients/product/__init__.py
@@ -1,0 +1,3 @@
+from ._product_client import ProductClient
+
+# flake8: noqa
