+++ nisystemlink/clients/product/models/_create_products_partial_success.py
@@ -2,7 +2,7 @@ from typing import List, Optional
 
 from nisystemlink.clients.core import ApiError
 from nisystemlink.clients.core._uplink._json_model import JsonModel
-from nisystemlink.clients.testmonitor.models import Product
+from nisystemlink.clients.product.models import Product
 
 
 class CreateProductsPartialSuccess(JsonModel):
