+++ nisystemlink/clients/product/utilities.py
@@ -1,20 +1,18 @@
 from typing import List
 
-from nisystemlink.clients.testmonitor._test_monitor_client import TestMonitorClient
-from nisystemlink.clients.testmonitor.models._paged_products import PagedProducts
-from nisystemlink.clients.testmonitor.models._product import Product
-from nisystemlink.clients.testmonitor.models._query_products_request import (
+from nisystemlink.clients.product._product_client import ProductClient
+from nisystemlink.clients.product.models._paged_products import PagedProducts
+from nisystemlink.clients.product.models._product import Product
+from nisystemlink.clients.product.models._query_products_request import (
     QueryProductsRequest,
 )
 
 
-def get_products_linked_to_file(
-    client: TestMonitorClient, file_id: str
-) -> List[Product]:
+def get_products_linked_to_file(client: ProductClient, file_id: str) -> List[Product]:
     """Gets a list of all the products that are linked to the file.
 
     Args:
-        `client` : The `TestMonitorClient` to use for the request.
+        `client` : The `ProductClient` to use for the request.
         `file_id`: The id of the file to query links for.
 
     Returns:
