+++ tests/integration/product/test_product_client.py
@@ -3,24 +3,24 @@ from typing import List
 
 import pytest
 from nisystemlink.clients.core._http_configuration import HttpConfiguration
-from nisystemlink.clients.testmonitor import models, TestMonitorClient
-from nisystemlink.clients.testmonitor.models import (
+from nisystemlink.clients.product._product_client import ProductClient
+from nisystemlink.clients.product.models import (
     CreateProductsPartialSuccess,
     Product,
 )
-from nisystemlink.clients.testmonitor.models._paged_products import PagedProducts
-from nisystemlink.clients.testmonitor.models._query_products_request import (
+from nisystemlink.clients.product.models._paged_products import PagedProducts
+from nisystemlink.clients.product.models._query_products_request import (
     ProductField,
     QueryProductsRequest,
     QueryProductValuesRequest,
 )
-from nisystemlink.clients.testmonitor.utilities import get_products_linked_to_file
+from nisystemlink.clients.product.utilities import get_products_linked_to_file
 
 
 @pytest.fixture(scope="class")
-def client(enterprise_config: HttpConfiguration) -> TestMonitorClient:
-    """Fixture to create a TestMonitorClient instance."""
-    return TestMonitorClient(enterprise_config)
+def client(enterprise_config: HttpConfiguration) -> ProductClient:
+    """Fixture to create a ProductClient instance."""
+    return ProductClient(enterprise_config)
 
 
 @pytest.fixture
@@ -31,7 +31,7 @@ def unique_identifier() -> str:
 
 
 @pytest.fixture
-def create_products(client: TestMonitorClient):
+def create_products(client: ProductClient):
     """Fixture to return a factory that creates specs."""
     responses: List[CreateProductsPartialSuccess] = []
 
@@ -51,13 +51,10 @@ def create_products(client: TestMonitorClient):
 
 @pytest.mark.integration
 @pytest.mark.enterprise
-class TestTestMonitor:
-    def test__api_info__returns(self, client: TestMonitorClient):
-        response = client.api_info()
-        assert len(response.dict()) != 0
+class TestProductClient:
 
     def test__create_single_product__one_product_created_with_right_field_values(
-        self, client: TestMonitorClient, create_products, unique_identifier
+        self, client: ProductClient, create_products, unique_identifier
     ):
         part_number = unique_identifier
         name = "Test Name"
@@ -83,7 +80,7 @@ class TestTestMonitor:
         assert created_product.properties == properties
 
     def test__create_multiple_products__multiple_creates_succeed(
-        self, client: TestMonitorClient, create_products
+        self, client: ProductClient, create_products
     ):
         products = [
             Product(part_number=uuid.uuid1().hex),
@@ -94,7 +91,7 @@ class TestTestMonitor:
         assert len(response.products) == 2
 
     def test__create_single_product_and_get_products__at_least_one_product_exists(
-        self, client: TestMonitorClient, create_products, unique_identifier
+        self, client: ProductClient, create_products, unique_identifier
     ):
         products = [Product(part_number=unique_identifier)]
         create_products(products)
@@ -103,7 +100,7 @@ class TestTestMonitor:
         assert len(get_response.products) >= 1
 
     def test__create_multiple_products_and_get_products_with_take__only_take_returned(
-        self, client: TestMonitorClient, create_products, unique_identifier
+        self, client: ProductClient, create_products, unique_identifier
     ):
         products = [
             Product(part_number=unique_identifier),
@@ -115,19 +112,19 @@ class TestTestMonitor:
         assert len(get_response.products) == 1
 
     def test__create_multiple_products_and_get_products_with_count_at_least_one_count(
-        self, client: TestMonitorClient, create_products, unique_identifier
+        self, client: ProductClient, create_products, unique_identifier
     ):
         products = [
             Product(part_number=unique_identifier),
             Product(part_number=unique_identifier),
         ]
         create_products(products)
-        get_response: models.PagedProducts = client.get_products(return_count=True)
+        get_response: PagedProducts = client.get_products(return_count=True)
         assert get_response is not None
         assert get_response.total_count is not None and get_response.total_count >= 2
 
     def test__get_product_by_id__product_matches_expected(
-        self, client: TestMonitorClient, create_products, unique_identifier
+        self, client: ProductClient, create_products, unique_identifier
     ):
         part_number = unique_identifier
         products = [Product(part_number=part_number)]
@@ -139,7 +136,7 @@ class TestTestMonitor:
         assert product.part_number == part_number
 
     def test__query_product_by_part_number__matches_expected(
-        self, client: TestMonitorClient, create_products, unique_identifier
+        self, client: ProductClient, create_products, unique_identifier
     ):
         part_number = unique_identifier
         products = [Product(part_number=part_number)]
@@ -153,7 +150,7 @@ class TestTestMonitor:
         assert query_response.products[0].part_number == part_number
 
     def test__query_product_values_for_name__name_matches(
-        self, client: TestMonitorClient, create_products, unique_identifier
+        self, client: ProductClient, create_products, unique_identifier
     ):
         part_number = unique_identifier
         test_name = "query values test"
@@ -170,7 +167,7 @@ class TestTestMonitor:
         assert query_response[0] == test_name
 
     def test__update_keywords_with_replace__keywords_replaced(
-        self, client: TestMonitorClient, create_products, unique_identifier
+        self, client: ProductClient, create_products, unique_identifier
     ):
         original_keyword = "originalKeyword"
         updated_keyword = "updatedKeyword"
@@ -191,7 +188,7 @@ class TestTestMonitor:
         assert original_keyword not in update_response.products[0].keywords
 
     def test__update_keywords_no_replace__keywords_appended(
-        self, client: TestMonitorClient, create_products, unique_identifier
+        self, client: ProductClient, create_products, unique_identifier
     ):
         original_keyword = "originalKeyword"
         additional_keyword = "additionalKeyword"
@@ -215,7 +212,7 @@ class TestTestMonitor:
         )
 
     def test__update_properties_with_replace__properties_replaced(
-        self, client: TestMonitorClient, create_products, unique_identifier
+        self, client: ProductClient, create_products, unique_identifier
     ):
         new_key = "newKey"
         original_properties = {"originalKey": "originalValue"}
@@ -240,7 +237,7 @@ class TestTestMonitor:
         )
 
     def test__update_properties_append__properties_appended(
-        self, client: TestMonitorClient, create_products, unique_identifier
+        self, client: ProductClient, create_products, unique_identifier
     ):
         original_key = "originalKey"
         new_key = "newKey"
@@ -270,7 +267,7 @@ class TestTestMonitor:
         assert updated_product.properties[new_key] == new_properties[new_key]
 
     def test__query_products_linked_to_files_correct_products_returned(
-        self, client: TestMonitorClient, create_products
+        self, client: ProductClient, create_products
     ):
         file_id = uuid.uuid1().hex
         product_name_with_file = "Has File"
