+++ tests/integration/testmonitor/test_testmonitor.py
@@ -1,0 +1,17 @@
+import pytest
+from nisystemlink.clients.core._http_configuration import HttpConfiguration
+from nisystemlink.clients.testmonitor import TestMonitorClient
+
+
+@pytest.fixture(scope="class")
+def client(enterprise_config: HttpConfiguration) -> TestMonitorClient:
+    """Fixture to create a TestMonitorClient instance."""
+    return TestMonitorClient(enterprise_config)
+
+
+@pytest.mark.integration
+@pytest.mark.enterprise
+class TestTestMonitor:
+    def test__api_info__returns(self, client: TestMonitorClient):
+        response = client.api_info()
+        assert len(response.dict()) != 0
